refactor(auth): update import paths for auth schemas and hooks
- Changed import paths in `auth.schema.test.ts` to directly reference `auth.schema`.
- Updated imports in `hooks.test.tsx` and `hooks.ts` to point to the correct context and hooks files, improving clarity and organization.

This refactor enhances the structure of the authentication feature by ensuring that imports are correctly aligned with the file organization.
